requirements.txt
# Web API
fastapi
uvicorn
python-multipart
python-dotenv

# AI / RAG
langchain
langchain-community
langchain-google-genai
google-genai
chromadb
pypdf

# Voice (Cloud Services)
groq
sarvamai

# Utilities
pydantic
requests
httpx
